refactored people controller and added people service
