+++ src/users/entities/user.entity.ts
@@ -20,7 +20,7 @@ const hashCost = 10;
 @Entity()
 @Unique(['username'])
 export class User {
-  @PrimaryGeneratedColumn()
+  @PrimaryColumn()
   id: number;
 
   @Column({ length: 40 })
